#1894 - validate create args
